WIP elevation tutorial
https://catlikecoding.com/unity/tutorials/hex-map/part-3/

Got to here: "Add a slider to the canvas via GameObject / Create / Slider"

The slider is being created but at the bottom where the game grid is rather than the expected spot on the editor ui. it does appear tho so that's a win! i've also not figured out how to get it to accept hexmapeditor's setelevation function as the "thing" it's supposed to be doing.

I also looked at this https://www.youtube.com/watch?v=nTLgzvklgU8 as a brief source which let me know that it has to be created from the "scene" panel, not as an element of the color panel.

Scene is also not generating correctly because elevation isn't done etc etc so don't panic about that :S the prior commit is the last working one.

edit: deleted scene changes, they were busted and breaking everything