elevation working!
I had to add and remove the hexmapeditor object a few times from the slider along with setting the function.  The issue was that the slider on value changed was setting a default value of 0 no matter what the slider position was. adding and removing the hexmadeditor object and set elevation function eventually removed the input value with the default 0. after that elevation Just Worked.

edit: it turns out that the above is just cargo cult. I added and removed it a few more times and the number returned and was set to zero but it still worked. WTF!

Anyway I stopped here https://catlikecoding.com/unity/tutorials/hex-map/part-3/

Repositioning Cell Labels